+++ src/main/java/com/commercetools/sync/producttypes/helpers/AttributeDefinitionCustomBuilder.java
@@ -7,14 +7,16 @@ import io.sphere.sdk.products.attributes.AttributeDefinitionBuilder;
 import io.sphere.sdk.products.attributes.AttributeDefinitionDraft;
 
 import javax.annotation.Nonnull;
-import java.util.Optional;
+
+import static java.util.Optional.ofNullable;
 
 
 public final class AttributeDefinitionCustomBuilder {
     /**
      * Given an {@link AttributeDefinitionDraft} returns a {@link AttributeDefinition} as a result. If the
      * 'attributeConstraint' field is 'null', its default value will be 'NONE'. If the 'inputHint' field is
-     * 'null, its default value will be 'SINGLE_LINE'.
+     * 'null', its default value will be 'SINGLE_LINE'. If the 'isRequired' field is null, its default value will be
+     * 'false'. If the 'isSearchable' field is null, its default value will be 'false'.
      *
      * @param attributeDefinitionDraft  the attribute definition draft which should be transformed.
      * @return The attribute definition with the same fields as the attribute definition draft.
@@ -25,15 +27,19 @@ public final class AttributeDefinitionCustomBuilder {
                 attributeDefinitionDraft.getName(),
                 attributeDefinitionDraft.getLabel(),
                 attributeDefinitionDraft.getAttributeType())
-            .isRequired(attributeDefinitionDraft.isRequired())
-            .attributeConstraint(Optional
-                .ofNullable(attributeDefinitionDraft.getAttributeConstraint())
+            .isRequired(
+                ofNullable(attributeDefinitionDraft.isRequired())
+                .orElse(false))
+            .attributeConstraint(
+                ofNullable(attributeDefinitionDraft.getAttributeConstraint())
                 .orElse(AttributeConstraint.NONE)) // Default value is NONE according to commercetools API
             .inputTip(attributeDefinitionDraft.getInputTip())
-            .inputHint(Optional
-                .ofNullable(attributeDefinitionDraft.getInputHint())
+            .inputHint(
+                ofNullable(attributeDefinitionDraft.getInputHint())
                 .orElse(TextInputHint.SINGLE_LINE)) // Default value is SINGLE_LINE according to commercetools API
-            .isSearchable(attributeDefinitionDraft.isSearchable())
+            .isSearchable(
+                ofNullable(attributeDefinitionDraft.isSearchable())
+                .orElse(false)) // Default value is false according to commercetools API
             .build();
     }