+++ src/main/java/com/commercetools/sync/services/impl/ProductTypeServiceImpl.java
@@ -11,7 +11,6 @@ import io.sphere.sdk.producttypes.commands.ProductTypeCreateCommand;
 import io.sphere.sdk.producttypes.commands.ProductTypeUpdateCommand;
 import io.sphere.sdk.producttypes.queries.ProductTypeQuery;
 import io.sphere.sdk.producttypes.queries.ProductTypeQueryBuilder;
-import io.sphere.sdk.queries.QueryExecutionUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import javax.annotation.Nonnull;
@@ -25,6 +24,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static java.lang.String.format;
@@ -107,12 +107,13 @@ public class ProductTypeServiceImpl
             .plusPredicates(queryModel -> queryModel.key().isIn(keys))
             .build();
 
-
-        return QueryExecutionUtils.queryAll(syncOptions.getCtpClient(), productTypeQuery)
-                                  .thenApply(productTypes -> productTypes
-                                      .stream()
-                                      .peek(productType -> keyToIdCache.put(productType.getKey(), productType.getId()))
-                                      .collect(toSet()));
+        return CtpQueryUtils
+            .queryAll(syncOptions.getCtpClient(), productTypeQuery, Function.identity())
+            .thenApply(productTypes -> productTypes
+                .stream()
+                .flatMap(List::stream)
+                .peek(productType -> keyToIdCache.put(productType.getKey(), productType.getId()))
+                .collect(toSet()));
     }
 
     @Nonnull